+++ src/app/shared/services/scholarship/scholarship.service.ts
