+++ src/app/shared/services/scholarship/scholarship.ts
